+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * change: Mapx max\* and min\* method rename to max\*OrNull and min\*OrNull
 * change: Collectionx max\* and min\* method rename to max\*OrNull and min\*OrNull
 * change: Arrayx max\* and min\* method rename to max\*OrNull and min\*OrNull
+* new: Collectionx added emptyList, flatten method
 
 ## v1.0.0-beta02
 improve: Improve code